project1
1. Write a procedure:
a) drawPyramid(int n) which takes as an input one integer value n and then
output on console a pyramid as on figure below (for example for n=4):
...X...
..XXX..
.XXXXX.
XXXXXXX
b) drawChristmasTree(int n) which takes as an input one integer value n and
then output on console a Christmas tree in which last part height equals n. The tree consists of
pyramids of heights from 1 to n. The shape have to be as presented below (for n=4):
...X...
...X...
..XXX..
...X...
..XXX..
.XXXXX.
...X...
..XXX..
.XXXXX.
XXXXXXX
2. Write a procedure loadDocument(String name) which will load and analyze lines
after lines searching for link in every line. The link format is as follows: 5 characters “link=”
after which the is a correct identifier. The correct identifier starts from letter (small or capital)
follows by letters or digits or underline ‘_’. The procedure has to print for every line all correct
identifiers in a separated line. Before printing, the identifiers have to be changed to small letters.
The document ends with line with the text “eod”, which means end of document.If an input line starts from ‘#’ sign or a line is empty, the line have to be ignored.
Else the input line have to be copied to output line with exclamation mark before first character.
Then the proper operation have to be done.
If a line has a format:
py n
your program has to call drawPyramid(n). You can assume that 1<=n<=20.
If a line has a format:
ct n
your program has to call drawChristmasTree(n). You can assume that 1<=n<=20.
If a line has a format:
ld docName
your program has to call loadDocument(String docName).
If a line has a format:
ha
your program has to end the execution, writing as the last line “END OF EXECUTION”. Every
test ends with this line.
For example for a test file:
py 3
ct 3
ld qwert
nnothing is here
link=abc link=qWe link=asd
link= broken li nk=wrong link =not
link=ok123_23sd what is here link=12wRong asdad link=_what12
dfasfdsdfsd
and now start LINK=$2323 LiNk=Ok
eod
ha
the output has to be:
START
!py 3
..X..
.XXX.
XXXXX
!ct 3
..X..
..X..
.XXX...X..
.XXX.
XXXXX
!ld qwert
abc
qwe
asd
ok123_23sd
ok
!ha
END OF EXECUTION